Added import/export functionality for deployments
This commit makes it much easier to reproduce deployments based on prior ones.

Deployment export is based on the deployment summary page, but output format
is yml, and internal IDs are removed, in favor of puppetclass names and Staypuft::Role
names, in order to ensure that the format is portable across deployments and
across foreman instances (assuming Role/puppetclass names are the same).

Import pulls a previously exported deployment param set into an existing deployment object.